convert AsyncTask to corountine
rename

use coroutine

simplify it
