+++ library/build.gradle
@@ -42,6 +42,12 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.2.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
+    implementation "androidx.core:core-ktx:1.2.0"
+    implementation "androidx.collection:collection-ktx:1.1.0"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.4"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.4"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.2.0"
+
     testImplementation 'junit:junit:4.13'
 }
