+++ library/src/main/java/info/hannes/cvscanner/util/SaveCallback.kt
@@ -1,0 +1,7 @@
+package info.hannes.cvscanner.util
+
+interface SaveCallback {
+    fun onSaveTaskStarted()
+    fun onSaved(savedPath: String?)
+    fun onSaveFailed(error: Exception?)
+}
